# docs/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "auth.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Auth"
    },
    {
      "name": "Route"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/route/getDestination/{id}": {
      "get": {
        "operationId": "Route_GetDestinationAndPolyline",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbDestintationAndPolylineResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Route"
        ]
      }
    },
    "/v1/route/sendDestinationAndPolyline": {
      "post": {
        "operationId": "Route_SendDestinationAndPolyline",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbDestintationAndPolylineResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbDestintationAndPolylineRequest"
            }
          }
        ],
        "tags": [
          "Route"
        ]
      }
    },
    "/v1/route/signUp": {
      "post": {
        "operationId": "Auth_SignUp",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbUserRequest"
            }
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    }
  },
  "definitions": {
    "pbDestintationAndPolylineRequest": {
      "type": "object",
      "properties": {
        "destination": {
          "$ref": "#/definitions/pbPoint"
        },
        "encodedRoute": {
          "type": "string"
        }
      }
    },
    "pbDestintationAndPolylineResponse": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "integer",
          "format": "int32"
        },
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/pbDestintationAndPolylineType"
        }
      }
    },
    "pbDestintationAndPolylineType": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "encodedRoute": {
          "type": "string"
        },
        "destination": {
          "$ref": "#/definitions/pbPoint"
        }
      }
    },
    "pbPoint": {
      "type": "object",
      "properties": {
        "latitude": {
          "type": "number",
          "format": "double"
        },
        "longitude": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "pbUserRequest": {
      "type": "object",
      "properties": {
        "googleId": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "pbUserResponse": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "integer",
          "format": "int32"
        },
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/pbUserType"
        },
        "token": {
          "type": "string"
        }
      }
    },
    "pbUserType": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "googleId": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}